services/back/pyproject.toml
[tool.poetry]
name = "wl-back"
version = "0.1.0"
description = ""
authors = ["Sergeev O. V. <o.v.sergeyev@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.9"
fastapi = "^0.110.0"
uvicorn = "^0.30.0"
requests = "^2.32.2"
pydantic-settings = "^2.2.1"
gunicorn = "^22.0.0"
passlib = {extras = ["bcrypt"], version = "^1.7.4"}
pyjwt = "^2.8.0"
sqlalchemy = "^2.0.31"
asyncpg = "^0.29.0"
pydantic = {extras = ["email"], version = "^2.8.2"}
starlette = {extras = ["middleware-cookiemiddleware"], version = "^0.37.2"}
starlette-securecookies = "^1.1.1"
alembic = "^1.13.2"
yookassa = "^3.3.0"
pytz = "^2024.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
